+++ include/mbgl/style/expression/expression.hpp
@@ -179,7 +179,8 @@ enum class Kind : int32_t {
     In,
     Within,
     Distance,
-    IndexOf
+    IndexOf,
+    Slice
 };
 
 class Expression {