+++ src/mbgl/style/expression/parsing_context.cpp
@@ -26,6 +26,7 @@
 #include <mbgl/style/expression/step.hpp>
 #include <mbgl/style/expression/within.hpp>
 #include <mbgl/style/expression/index_of.hpp>
+#include <mbgl/style/expression/slice.hpp>
 
 #include <mbgl/style/expression/find_zoom_curve.hpp>
 #include <mbgl/style/expression/dsl.hpp>
@@ -141,6 +142,7 @@ MAPBOX_ETERNAL_CONSTEXPR const auto expressionRegistry =
         {"var", Var::parse},
         {"within", Within::parse},
         {"index-of", IndexOf::parse},
+        {"slice", Slice::parse}
     });
 
 bool isExpression(const std::string& name) {